Convert tests to TypeScript
